studenthub/requirements.txt
fastapi
uvicorn
motor
pydantic
bcrypt
cloudinary
pyjwt
python-multipart